Refactor mongo.js to add and retrieve persons
The functionality of mongo.js has been refactored to manage persons instead of notes. Now, the script accepts name and number as command line arguments to create a 'Person' in MongoDB. If these arguments are not supplied, the script fetches and prints all 'Persons' in the database. The operation has been encapsulated into different functions for clean and distinct functionality.